requirements.txt
PySide6>=6.4.0
Pillow>=9.3.0
pytesseract>=0.3.10
openai>=1.3.0
numpy>=1.24.0
python-dotenv>=1.0.0 